[157] remove some code duplication, remove console.logs and fix dropdown placeholder
