+++ src/Components/SectorDropdown/index.js
@@ -2,9 +2,17 @@ import React, { useEffect, useState } from 'react';
 import Select from 'react-select';
 import RightBoxInputs from './Style';
 import { Label } from '../UserDropdown/Style';
+import customStyles from './dropdownStyle';
 
 const SectorDropdown = ({ setSector, sectorID }) => {
   const [id, setId] = useState('');
+  const [placeholder, setPlaceholder] = useState('Setor');
+
+  useEffect(() => {
+    if (sectorID) {
+      setPlaceholder(sectorID);
+    }
+  }, [sectorID]);
   const options = [
     { value: 'DPSS', label: 'DPSS' },
     { value: 'Clinica', label: 'Clinica' },
@@ -14,36 +22,13 @@ const SectorDropdown = ({ setSector, sectorID }) => {
     setSector(id.label);
   }, [id]);
 
-  const customStyles = {
-    option: (provided) => ({
-      ...provided,
-      borderBottom: '1px dotted pink',
-      color: 'black',
-      padding: '5%',
-    }),
-    control: () => ({
-      color: 'white',
-      display: 'flex',
-      borderRadius: '10px',
-      border: '1px solid #FFFFFF',
-      width: '25vw',
-    }),
-    singleValue: (provided, state) => {
-      const opacity = state.isDisabled ? 0.5 : 1;
-      const transition = 'opacity 300ms';
-      const color = 'white';
-      return {
-        ...provided, opacity, transition, color,
-      };
-    },
-  };
   return (
     <RightBoxInputs>
       <Label>
         Setor:
       </Label>
       <Select
-        placeholder={sectorID}
+        placeholder={placeholder}
         styles={customStyles}
         options={options}
         onChange={(value) => setId(value)}
