+++ src/Components/UserDropdown/index.js
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from 'react';
 import Select from 'react-select';
 import { UserSearchDiv, Label } from './Style';
 import './input.css';
+import customStyles from '../SectorDropdown/dropdownStyle';
 
 const UserDropdown = ({ clients, setClientID, setClientName }) => {
   const [id, setId] = useState('');
@@ -14,30 +15,6 @@ const UserDropdown = ({ clients, setClientID, setClientName }) => {
     setClientID(id.value);
     setClientName(id.label);
   }, [id]);
-
-  const customStyles = {
-    option: (provided) => ({
-      ...provided,
-      borderBottom: '1px dotted pink',
-      color: 'black',
-      padding: '5%',
-    }),
-    control: () => ({
-      color: 'white',
-      display: 'flex',
-      borderRadius: '10px',
-      border: '1px solid #FFFFFF',
-      width: '25vw',
-    }),
-    singleValue: (provided, state) => {
-      const opacity = state.isDisabled ? 0.5 : 1;
-      const transition = 'opacity 300ms';
-      const color = 'white';
-      return {
-        ...provided, opacity, transition, color,
-      };
-    },
-  };
 
   return (
     <UserSearchDiv>
